+++ .dockerignore
@@ -1,0 +1,15 @@
+tests
+data
+.github
+__pycache__
+.gitignore
+README.md
+.pre-commit-config.yaml
+pyproject.toml
+requirements-dev.txt
+run_dev.py
+postgres-data
+.flaskenv
+.flask_secret
+.env
+*.env
