+++ .gitignore
@@ -14,3 +14,4 @@ venv/
 .pytest_cache
 .flask_secret
 .env
+*.env
