+++ app.env-example
@@ -1,0 +1,3 @@
+# environment variables for docker (app specific)
+DB_BACKEND=postgres
+POSTGRES_HOST=nlpdb
