+++ db.env-example
@@ -1,0 +1,5 @@
+# docker environment variables# environment variables for docker (database specific)
+POSTGRES_USER=nlp4all
+POSTGRES_PASSWORD=nlp4all
+POSTGRES_DB=nlp4all
+POSTGRES_PORT=5432
