--- docker-compose.yml
@@ -1,24 +1,0 @@
-version: '2.1'
-services:
-  postgres:
-    restart: always
-    image: postgres:15.1
-    environment:
-      - POSTGRES_USER=nlp4all
-      - POSTGRES_PASSWORD=nlp4all
-      - POSTGRES_DB=nlp4all
-    volumes:
-      - ./postgres-data/postgres:/var/lib/postgresql/data
-    ports:
-      - "5432:5432"
-  app:
-    restart: always
-    image: python:3.11
-    command: sh -c "pip install --no-cache-dir -r requirements.txt && python run.py"
-    working_dir: /app
-    volumes:
-      - ./:/app
-    ports:
-      - 5000:5000
-    environment:
-      - FLASK_ENV=development
