+++ docker/nginx/Dockerfile
@@ -1,0 +1,3 @@
+FROM nginx:1.23.2
+RUN rm /etc/nginx/conf.d/default.conf
+COPY ./nginx.conf /etc/nginx/nginx.conf
