+++ docker/nginx/nginx.conf
@@ -1,0 +1,34 @@
+
+user  nginx;
+worker_processes  auto;
+
+error_log  /var/log/nginx/error.log notice;
+pid        /var/run/nginx.pid;
+
+
+events {
+    worker_connections  1024;
+}
+
+
+http {
+    upstream nlp4all_app {
+        server app:8000;
+    }
+
+    server {
+        listen 80;
+
+        location / {
+            proxy_pass http://nlp4all_app;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header Host $host;
+            proxy_redirect off;
+        }
+
+        location /static/ {
+            alias /home/app/nlp4all/static/;
+        }
+
+    }
+}
