+++ docker/python/entrypoint.sh
@@ -1,0 +1,14 @@
+#!/bin/sh
+
+if [ "$DB_BACKEND" = "postgres" ]
+then
+    echo "Waiting for postgres..."
+
+    while ! nc -z $POSTGRES_HOST $POSTGRES_PORT; do
+      sleep 0.1
+    done
+
+    echo "PostgreSQL started"
+fi
+
+exec "$@"
