--- init-db.sh
@@ -1,4 +1,0 @@
-clear
-python dbinit.py
-python run.py
-
