+++ nlp4all/__init__.py
@@ -3,25 +3,23 @@ nlp4all module
 """
 
 import os
-from flask import Flask
+
+from flask import Flask, send_from_directory
 from flask_login import LoginManager
 from flask_cors import CORS
-from flask_bcrypt import Bcrypt
+# from flask_bcrypt import Bcrypt
+from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
+
+from .helpers import database as dbhelper
 from .config import get_config, Config
-from .models import database
-from .models import User
+from .models.database import Base
+from .models import load_user
 from .routes import Router
 
-def load_user(user_id):
-        """Loads a user from the database.
+db = SQLAlchemy(model_class = Base)
+migrate = Migrate()
 
-        Args:
-            user_id (int): The id of the user to load.
-
-        Returns:
-        User: User object.
-        """
-        return User.query.get(int(user_id))
 
 def create_app(env: str = "production") -> Flask:
     """Create the Flask app."""
@@ -31,8 +29,9 @@ def create_app(env: str = "production") -> Flask:
     app.secret_key = conf.get_secret()
     app.config.from_object(conf)
     os.environ.setdefault("SQLALCHEMY_DATABASE_URI", conf.SQLALCHEMY_DATABASE_URI)
-    
-    database.init_app(app)
+
+    db.init_app(app)
+    migrate.init_app(app, db)
 
     Router.run(app)
 
@@ -43,8 +42,16 @@ def create_app(env: str = "production") -> Flask:
     login_manager.user_loader(load_user)
     login_manager.login_message_category = "info"
 
+    dbhelper.init_app(app)
+
+    if env != "production":
+        from .helpers import development # pylint: disable=wrong-import-position
+        app.add_url_rule('/api/help', methods = ['GET'], view_func=development.help_route)
+    
+    
+    @app.route("/static/<path:filename>")
+    def staticfiles(filename):
+        """Static file router"""
+        return send_from_directory(app.config["STATIC_DIR"], filename)
+
     return app
-
-
-
-
