+++ requirements.txt
@@ -24,6 +24,7 @@ Flask-Email==1.4.4
 Flask-Login==0.6.2
 Flask-Mail==0.9.1
 flask_migrate==4.0.0
+flask_sqlalchemy==3.0.2
 Flask-WTF==1.0.1
 greenlet==2.0.1
 gunicorn==20.1.0
@@ -51,6 +52,7 @@ pip-upgrader==1.4.15
 platformdirs==2.5.3
 pluggy==1.0.0
 pre-commit==2.20.0
+psycopg2==2.9.5
 pycparser==2.21
 PyJWT==2.6.0
 PySocks==1.7.1
