--- run.py
@@ -1,7 +1,0 @@
-"""
-nlp4all flask app runner.
-"""
-from nlp4all import app
-
-if __name__ == "__main__":
-    app.run(debug=False)
